fix: some endpoints moved to admin controller
