+++ src/main/resources/application-dev.yaml
@@ -5,14 +5,14 @@ jwt:
 spring:
   datasource:
     driver-class-name: org.postgresql.Driver
-    url: jdbc:postgresql://localhost:5432/teamChallenge
+    url: jdbc:postgresql://localhost:32768/teamChallenge
     username: postgres
-    password: 123
+    password: 1234
 
   jpa:
     hibernate:
       ddl-auto: update
-    defer-datasource-initialization: false
+    defer-datasource-initialization: true
   liquibase:
     change-log: classpath:db/liquibase-changelog.xml
